+++ src/Libraries/SmartStore.Core/Domain/Payments/RestrictionContextTypes.cs
@@ -1,0 +1,15 @@
+﻿
+namespace SmartStore.Core.Domain.Payments
+{
+	public enum CountryExclusionContextType
+	{
+		BillingAddress = 0,
+		ShippingAddress
+	}
+
+	public enum AmountRestrictionContextType
+	{
+		SubtotalAmount = 0,
+		TotalAmount
+	}
+}