+++ src/Tests/SmartStore.Core.Tests/Domain/Shipping/ShippingOptionListTypeConverterTests.cs
@@ -28,6 +28,7 @@ namespace SmartStore.Core.Tests.Domain.Shipping
             var shippingOptionsInput = new List<ShippingOption>();
             shippingOptionsInput.Add(new ShippingOption()
             {
+				ShippingMethodId = 1,
                 Name = "a1",
                 Description = "a2",
                 Rate = 3.57M,
@@ -35,6 +36,7 @@ namespace SmartStore.Core.Tests.Domain.Shipping
             });
             shippingOptionsInput.Add(new ShippingOption()
             {
+				ShippingMethodId = 2,
                 Name = "b1",
                 Description = "b2",
                 Rate = 7.00M,
@@ -47,11 +49,13 @@ namespace SmartStore.Core.Tests.Domain.Shipping
             var shippingOptionsOutput = converter.ConvertFrom(result) as List<ShippingOption>;
             shippingOptionsOutput.ShouldNotBeNull();
             shippingOptionsOutput.Count.ShouldEqual(2);
+			shippingOptionsOutput[0].ShippingMethodId.ShouldEqual(1);
             shippingOptionsOutput[0].Name.ShouldEqual("a1");
             shippingOptionsOutput[0].Description.ShouldEqual("a2");
             shippingOptionsOutput[0].Rate.ShouldEqual(3.57M);
             shippingOptionsOutput[0].ShippingRateComputationMethodSystemName.ShouldEqual("a4");
 
+			shippingOptionsOutput[1].ShippingMethodId.ShouldEqual(2);
             shippingOptionsOutput[1].Name.ShouldEqual("b1");
             shippingOptionsOutput[1].Description.ShouldEqual("b2");
             shippingOptionsOutput[1].Rate.ShouldEqual(7.00M);